Works: print out things along side other actions

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/configs/ConfigManager.java
@@ -17,7 +17,7 @@ public class ConfigManager {
         registerField("intakeActive", IntakeSetting.class, new IntakeSetting("NORMAL", true, 0.8));
         registerField("motorName", String.class, "motorName");
         registerField("gamepadLeftY", Double.class, 0.0);
-        registerField("sysTime", Integer.class, 0);
+        registerField("sysTime", Long.class, 0L);
 
         // New required fields for Auto
         registerField("Title", String.class, "");
